vercel-issue-resolved

+++ src/app/api/deleteslot/[id]/route.ts
@@ -1,18 +1,23 @@
 import { connectDB } from "@/app/lib/connectDB";
 import { NextRequest, NextResponse } from "next/server";
 import Slot from "@/app/model/slotSchema";
-// POST request handler
+
 export async function DELETE(
   req: NextRequest,
-  { params }: { params: { id: string } }
+  context: { params: Record<string, string> }
 ) {
   try {
     await connectDB();
-    const { id } = params;
+
+    const id = context.params.id;
+    if (!id) {
+      return NextResponse.json({ error: "ID is required" }, { status: 400 });
+    }
 
     const slot = await Slot.findByIdAndDelete(id);
+
     return NextResponse.json(
-      { message: "Slots deleted successfully", slot },
+      { message: "Slot deleted successfully", slot },
       { status: 200 }
     );
   } catch (err: unknown) {
